fix: factor out notifications filters
